Use protocol-independent paths to submodules
With paths like `../../NilFoundation/` submodules will be cloned with
the same protocol as used to clone the parent repository.

Unlike explicit SSH, it allows external contributors and CI servers
to clone this repository without having to use SSH keys.

Unlike explicit HTTPS, it allows our own developers to develop code
in submodules and push it to GitHub without having to enter usernames
and passwords every time or to store tokens on development machines.

Resolves #33