ci: Clean index.lock files
Cancelling a job while the checkout step is running often results in
index.lock files persisting in one of the submodules. The checkout action
only removes index.lock files in the root repo, but not in submodules.
(See issue https://github.com/actions/checkout/issues/1153).
On the following job runs, checkout fails, being unable to operate on
a submodule with an index.lock file.

This new step will remove such files when the job was cancelled or
has failed on the checkout step or before it.

For details on the status check functions, used in this step, see
https://docs.github.com/en/actions/learn-github-actions/expressions#status-check-functions

Resolves #316

Co-authored-by: Mikhail Aksenov <maksenov@nil.foundation>